Palindromes
A palindromes is a sequence of characters that reads the same backward
as forward. This is an app that reads in a five-digit integer and
determines whether it is a palindrome. If the number is not five digits
long, display an error message and allow the user to enter a new value.